permutations: Add property based checks
This commit adds basic properties for the functions found in the module
permutations.cry.

The choice of checking powers of two up to 2^^4096 is arbitrary --- it
seems plenty high and was still a fast check.

The choice of 256 bits as the size of values in the other two checks is
based on the fact that elements in the field for the Python spec are
256-bits.

In order to formally prove these properties, load the
permutations_properties.cry module into Cryptol and then use the "prove"
command. For example:

   $> cryptol permutations_properties.cry
   Main> prove: check_bit_reversal_permutation_twice_is_value

gives the output:

   Q.E.D.
   (Total Elapsed Time: 0.266s, using "Z3")