reverse_bits: Rename type parameter for clarity
Prior to this commit, the type parameter name of `order` in the function
`reverse_bits` was confusing, because it did not match the Python spec.

As noted in a code review:
"The spec has order as the power of two (2^m) but this implementation
has order as the bit length of that, which is m."

Thus I renamed the type parameter to `order_bit_length` to clarify this
distinction and deconflict from the Python specification use of order.