proof: Switch permutations to use submodule source code
With the c-kzg source code now included as a submodule, we want our
proof script to use it. This commit does just that by:
 1) Adding a Makefile for generating the necessary LLVM bitcode from
    `ckzg.c` with instructions found in the README
 2) Updating `permutations.saw` to point to the generated bitcode

Run `make bc` inside `proof/bitcode` to generate the `ckzg.bc` file
needed by the proofs in `permutations.saw`.

We had to modify the Makefile inside the submodule, `c-kzg-4844/src/`,
to add a new command that emits LLVM bitcode. But we don't want to fork
the c-kzg repository. So we created a patch with the necessary updates
that gets applied when running the Makefile found in `proof/bitcode/`.