bitcode: Fix error in make bc
As introduced in a previous commit, `make bc` generated the need LLVM
bitcode file, `ckzg.bc` from the source in the c-kzg submodule, and
copied it into the `/proof` folder. This only worked on the first call,
with proceeding calls failing to apply the Git patch cleanly because it
had already been applied.

This commit fixes the Makefile `bc` command so that it doesn't try to
re-apply the patch file, thus fixing this error.