spec: Refactor modules into separate folders
The new hierarchy is:
 - spec/Spec/Permutations.cry
 - spec/Common/Utils.cry

This is in preparation to add another subfolder, LLVM, to hold Cryptol
specs that more closely match the structure of the C code we are trying
to formally verify.

Now that we don't have a flat folder hierarchy for all Cryptol modules,
we need to set the enviroment variable CRYPTOLPATH to tell Cryptol where
to look for dependencies. Thus this commit updates the Makefile in the
'proof' folder so that it will set the CRYPTOLPATH to 'spec' for all
runs of SAW.