util: Switch modular inverse and div to use Integer type
I switched the API to use the Integer type for these utility functions
because it simplifies the code. The BlsHelper functions that call these
utility functions have been updated and now handle the conversion of
types (Integral <=> Integer) internally.

This also exposed an error in the modular_inverse when a negative
number was passed. A new test has been added to make sure we don't
regress.