utils: Replace uses of Integer with Z to address overflow concerns
This commit addresses several issues raised during code review. Namely,
the Cryptol specs in Utils, BlsEC, and BlsHelpers modules are 1) overly
complex by using the Integer type for modular arithmetic and 2) have
potential overflow issues when using Integers and fixed width size
types.

I was not previously aware of the type Z in Cryptol, which is a built-in
type for integer fields module n, specified as (Z n), for some constant
modulus n.

By replacing the usage of Integer with Z we will be able to reduce the
complexity of the specs and eliminate potential overflow errors. Also,
using Z will actually match the Python specification better, since it
uses the py_ecc type FQ, a Python class for modular field elements and
operations.

Next steps:
 - Replace the Integer type to the Z type in the functions of
   the BlsEC and BlsHelpers modules (this will take more refactoring).