utils: Rename div and pow functions with _mod postfix for clarity
My previous thinking was to have these function names match the Python
equivalents, but it came up in code review that adding the _mod postfix
would improve clarity for anyone that needs to look at the code in the
future.

As I thought more about this, I see how the Python code makes it
seem like these two functions are built-in, but the actual
implementations are overloaded for a specific arithmetic field object
type. I want to avoid this small "gotch ya."