blsec: Remove Fp value and rename FP type to Fp
After the previous changes, I noted that the value Fp was only used in
one place, and I could see no benefit to using Fp in place of `(FP).
So, I decided to replace that instance and remove the value Fp.

I prefer the name Fp over FP, so now that only the type FP is used,
I renamed the type FP to Fp.