blsserde: Remove polymorphic type n from get_flags
And hardcode to work on a 384-bit value. I'm doing this for two reasons:
first, because it better matches the Python spec; second, because it
will only ever be used a 384-bit value, so there is no need for it to be
more abstract; and third, because it reduces complexity making the
function be monomorphic.