blshelpers: Fix error in compute_powers
Not sure how I missed this previously, but I started the array of powers
with x^1 instead of x^0.

I've fixed the code, added a new test --- that matches the Python spec;
also removed the test_compute_powers in favor of the doc-string REPL
test. Finally, I also simplified the indexing of the list comprehension
in compute_powers.