spec: Replace :prove with :check when test has no symbolic values
In all of the tests cases where there are concrete inputs and outputs,
there is no benefit to calling :prove and passing the computation to the
SMT solver, so instead we just run :check. This saves a slight amount of
runtime overhead, but more importantly it helps communicate what are
tests running explicit inputs compared to properties being proven for
all possible values of an input.